+++ lib/dom/getBody/index.js
@@ -1,0 +1,18 @@
+import {
+	document,
+	Promise
+} from 'j0';
+import wrap from '../wrap';
+
+export default new Promise(function (resolve) {
+	const interval = 50;
+	function check() {
+		const {body} = document;
+		if (body) {
+			resolve(wrap(body));
+		} else {
+			setTimeout(check, interval);
+		}
+	}
+	setTimeout(check);
+});
