+++ lib/dom/wrap/index.js
@@ -1,0 +1,7 @@
+import J0Element from '../J0Element';
+
+function wrap(source) {
+	return new J0Element(source);
+}
+
+export default wrap;
